# docs/swagger.yaml
basePath: /
definitions:
...
  buildings.BuildingListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/buildings.BuildingResponse'
        type: array
      page:
        example: 1
        type: integer
      page_size:
        example: 20
        type: integer
      total:
        example: 5
        type: integer
    type: object
...
  buildings.CreateBuildingRequest:
    properties:
      address_line_1:
        example: 100 Main St
        maxLength: 255
        type: string
      address_line_2:
        example: Building A
        maxLength: 255
        type: string
      city:
        example: Austin
        maxLength: 120
        type: string
      country:
        example: US
        type: string
      name:
        example: Riverside Tower
        maxLength: 255
        minLength: 1
        type: string
      postal_code:
        example: "78701"
        maxLength: 32
        type: string
      state:
        example: TX
        maxLength: 120
        type: string
      status:
        enum:
        - active
        - inactive
        - archived
        example: active
        type: string
      timezone:
        example: America/Chicago
        maxLength: 64
        type: string
    required:
    - name
    type: object
...
  leases.CreateLeaseRequest:
    properties:
      currency:
        example: USD
        type: string
      deposit_amount_minor:
        minimum: 0
        type: integer
      end_date:
        type: string
      is_primary:
        type: boolean
      monthly_rent_amount_minor:
        minimum: 0
        type: integer
      start_date:
        type: string
      status:
        enum:
        - draft
        - active
        - ended
        - cancelled
        example: draft
        type: string
      tenant_id:
        type: string
      unit_id:
        type: string
    required:
    - monthly_rent_amount_minor
    - start_date
    - tenant_id
    - unit_id
    type: object
...
  organizations.CreateOrganizationRequest:
    properties:
      address_line1:
        example: 1 Main St
        maxLength: 255
        type: string
      address_line2:
        example: Suite 200
        maxLength: 255
        type: string
      billing_email:
        example: billing@acme.com
        maxLength: 320
        type: string
      city:
        example: Austin
        maxLength: 120
        type: string
      company_registration:
        example: DE-12345
        maxLength: 128
        type: string
      country:
        example: US
        type: string
      default_currency:
        example: USD
        type: string
      legal_name:
        example: Acme Property Management LLC
        maxLength: 255
        type: string
      name:
        example: Acme Property Management
        maxLength: 255
        minLength: 2
        type: string
      phone:
        example: +1-555-0100
        maxLength: 64
        type: string
      postal_code:
        example: "78701"
        maxLength: 32
        type: string
      region:
        example: TX
        maxLength: 120
        type: string
      slug:
        example: acme-prop
        maxLength: 128
        minLength: 2
        type: string
      tax_id:
        example: 12-3456789
        maxLength: 64
        type: string
      timezone:
        example: America/Chicago
        maxLength: 64
        type: string
      website:
        example: https://acme.com
        maxLength: 512
        type: string
    required:
    - name
    - slug
    type: object
  organizations.OrganizationResponse:
    properties:
      address_line1:
        type: string
      address_line2:
        type: string
      billing_email:
        type: string
      city:
        type: string
      company_registration:
        type: string
      country:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      default_currency:
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      legal_name:
        type: string
      name:
        example: Acme Property Management
        type: string
      phone:
        type: string
      postal_code:
        type: string
      region:
        type: string
      slug:
        example: acme-prop
        type: string
      tax_id:
        type: string
      timezone:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      website:
        type: string
    type: object
  organizations.PatchOrganizationRequest:
    properties:
      address_line1:
        maxLength: 255
        type: string
      address_line2:
        maxLength: 255
        type: string
      billing_email:
        maxLength: 320
        type: string
      city:
        maxLength: 120
        type: string
      company_registration:
        maxLength: 128
        type: string
      country:
        type: string
      default_currency:
        type: string
      legal_name:
        maxLength: 255
        type: string
      name:
        maxLength: 255
        minLength: 2
        type: string
      phone:
        maxLength: 64
        type: string
      postal_code:
        maxLength: 32
        type: string
      region:
        maxLength: 120
        type: string
      tax_id:
        maxLength: 64
        type: string
      timezone:
        maxLength: 64
        type: string
      website:
        maxLength: 512
        type: string
    type: object
...
  response.Envelope-buildings_BuildingListResponse:
    properties:
      data:
        $ref: '#/definitions/buildings.BuildingListResponse'
    type: object
  response.Envelope-buildings_BuildingResponse:
    properties:
      data:
        $ref: '#/definitions/buildings.BuildingResponse'
    type: object
  response.Envelope-leases_LeaseListResponse:
    properties:
      data:
        $ref: '#/definitions/leases.LeaseListResponse'
    type: object
  response.Envelope-leases_LeaseResponse:
    properties:
      data:
        $ref: '#/definitions/leases.LeaseResponse'
    type: object
  response.Envelope-organizations_OrganizationResponse:
    properties:
      data:
        $ref: '#/definitions/organizations.OrganizationResponse'
    type: object
  response.Envelope-tenants_TenantListResponse:
    properties:
      data:
        $ref: '#/definitions/tenants.TenantListResponse'
    type: object
  response.Envelope-tenants_TenantResponse:
    properties:
      data:
        $ref: '#/definitions/tenants.TenantResponse'
    type: object
  response.Envelope-units_UnitListResponse:
    properties:
      data:
        $ref: '#/definitions/units.UnitListResponse'
    type: object
  response.Envelope-units_UnitResponse:
    properties:
      data:
        $ref: '#/definitions/units.UnitResponse'
    type: object
  response.Envelope-users_UserListResponse:
    properties:
      data:
        $ref: '#/definitions/users.UserListResponse'
    type: object
  response.Envelope-users_UserResponse:
    properties:
      data:
        $ref: '#/definitions/users.UserResponse'
    type: object
  response.ErrorBody:
    properties:
      error:
        $ref: '#/definitions/response.ErrorDetail'
    type: object
  response.ErrorDetail:
    properties:
      code:
        type: string
      kind:
        type: string
      message:
        type: string
    type: object
  tenants.CreateTenantRequest:
    properties:
      email:
        example: jordan.lee@example.com
        maxLength: 320
        type: string
      email_opt_in:
        example: true
        type: boolean
      first_name:
        example: Jordan
        maxLength: 120
        minLength: 1
        type: string
      full_name:
        example: Jordan Lee
        maxLength: 512
        type: string
      last_name:
        example: Lee
        maxLength: 120
        type: string
      locale:
        example: en-US
        maxLength: 32
        type: string
      phone:
        example: +1-555-0100
        maxLength: 32
        type: string
      sms_opt_in:
        example: false
        type: boolean
      sms_verified:
        example: false
        type: boolean
      status:
        enum:
        - active
        - inactive
        - archived
        example: active
        type: string
      timezone:
        example: America/Chicago
        maxLength: 64
        type: string
    required:
    - first_name
    type: object
...
  units.CreateUnitRequest:
    properties:
      bedrooms:
        example: 2
        maximum: 32
        minimum: 0
        type: integer
      currency:
        example: USD
        type: string
      default_rent_amount_minor:
        example: 185000
        minimum: 0
        type: integer
      status:
        enum:
        - vacant
        - occupied
        - offline
        - maintenance
        example: vacant
        type: string
      unit_label:
        example: 101-A
        maxLength: 64
        minLength: 1
        type: string
    required:
    - unit_label
    type: object
...
  users.CreateUserRequest:
    properties:
      email:
        example: alex@acme.com
        maxLength: 320
        type: string
      first_name:
        example: Alex
        maxLength: 120
        minLength: 1
        type: string
      last_name:
        example: Rivera
        maxLength: 120
        type: string
      password:
        example: SecurePass123
        maxLength: 72
        minLength: 8
        type: string
      phone:
        example: +1-555-0199
        maxLength: 64
        type: string
      role:
        enum:
        - admin
        - landlord
        - manager
        - staff
        example: staff
        type: string
      status:
        enum:
        - active
        - disabled
        - invited
        example: active
        type: string
    required:
    - email
    - first_name
    - role
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Rent Management System HTTP API (bootstrap; module routes are placeholders).
  title: RMS API
  version: "1.0"
paths:
...
  /api/v1/organizations:
    post:
      consumes:
      - application/json
      description: '**Admin only.** Creates a new landlord / property-management company
        (tenant). Sets `created_by` / `updated_by` from the access token subject.'
      parameters:
      - description: Organization payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/organizations.CreateOrganizationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Envelope-organizations_OrganizationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "403":
          description: Non-admin callers are rejected
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Create organization
      tags:
      - organizations
  /api/v1/organizations/{id}:
    get:
      description: '**Admin:** any organization id. **Landlord, manager, staff:**
        only their own `organization_id` from the token; cross-tenant access returns
        403.'
      parameters:
      - description: Organization ID
        example: 550e8400-e29b-41d4-a716-446655440000
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope-organizations_OrganizationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Get organization
      tags:
      - organizations
    patch:
      consumes:
      - application/json
      description: '**Admin:** may update any organization. **Landlord and manager:**
        may update only their own organization. **Staff** cannot use this endpoint.
        Slug is immutable. Sets `updated_by` from the token subject.'
      parameters:
      - description: Organization ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update (omit to leave unchanged)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/organizations.PatchOrganizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope-organizations_OrganizationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Update organization profile
      tags:
      - organizations
  /api/v1/organizations/{id}/buildings:
    get:
      description: '**Admin:** any organization `id`. **Landlord, manager, staff:**
        only buildings in their own organization. Paginated via `page` and `page_size`
        (max 100). Optional `status` filter: `active`, `inactive`, or `archived`.'
      parameters:
      - description: Organization ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - default: 1
        description: Page (1-based)
        example: 1
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Page size
        example: 20
        in: query
        maximum: 100
        name: page_size
        type: integer
      - description: 'Filter: active, inactive, or archived'
        example: active
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope-buildings_BuildingListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: List buildings
      tags:
      - buildings
    post:
      consumes:
      - application/json
      description: '**Admin:** create in any organization. **Landlord / manager:**
        create only in their own organization. **Staff** cannot create. Sets audit
        `created_by` / `updated_by` from the token subject.'
      parameters:
      - description: Organization ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: 'Example: {\'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/buildings.CreateBuildingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Envelope-buildings_BuildingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Create building
      tags:
      - buildings
...
  /api/v1/organizations/{id}/buildings/{buildingId}/units:
    get:
      description: '**Nested route:** `organizations` → `buildings` → `units`. **Admin:**
        any organization `id`. **Landlord, manager, staff:** only resources in their
        token organization. Paginated (`page`, `page_size`, max 100). Optional `status`
        filter (`vacant`, `occupied`, `offline`, `maintenance`). Set `include_occupancy=true`
        to add a derived `occupancy` object per unit (active lease summary when present).'
      parameters:
      - description: Organization ID (same wildcard as other org routes)
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Building ID (must belong to organization)
        format: uuid
        in: path
        name: buildingId
        required: true
        type: string
      - default: 1
        description: Page (1-based)
        example: 1
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Page size
        example: 20
        in: query
        maximum: 100
        name: page_size
        type: integer
      - description: Filter by unit status
        example: vacant
        in: query
        name: status
        type: string
      - description: Set to true/1/yes to include occupancy summary per unit
        example: "true"
        in: query
        name: include_occupancy
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Example items include unit_label, currency, default_rent_amount_minor;
            occupancy when requested
          schema:
            $ref: '#/definitions/response.Envelope-units_UnitListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: List units in a building
      tags:
      - units
    post:
      consumes:
      - application/json
      description: '**Admin / landlord / manager:** create a unit under the given
        building. **Staff** cannot create. `unit_label` must be unique per building.
        Defaults: `currency` USD, `status` vacant.'
      parameters:
      - description: Organization ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Building ID
        format: uuid
        in: path
        name: buildingId
        required: true
        type: string
      - description: 'Example: {\'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/units.CreateUnitRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Envelope-units_UnitResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Create unit
      tags:
      - units
...
  /api/v1/organizations/{id}/leases:
    get:
      description: Organization-scoped list with **pagination** (`page`, `page_size`,
        max 100). **Filters:** `status`, `building_id` (via unit), `unit_id`, `tenant_id`,
        `active_on` (date `YYYY-MM-DD` — leases **active** for any time on that UTC
        calendar day). **Overlap rule:** creating/updating an **active + primary**
        lease is rejected if its `[start_date,end_date)` range intersects another
        active primary lease on the same unit (mirrors DB EXCLUDE constraint).
      parameters:
      - description: Organization ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - default: 1
        description: Page (1-based)
        example: 1
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Page size
        example: 20
        in: query
        maximum: 100
        name: page_size
        type: integer
      - description: draft | active | ended | cancelled
        example: active
        in: query
        name: status
        type: string
      - description: Filter by building (units in this building)
        format: uuid
        in: query
        name: building_id
        type: string
      - description: Filter by unit
        format: uuid
        in: query
        name: unit_id
        type: string
      - description: Filter by tenant
        format: uuid
        in: query
        name: tenant_id
        type: string
      - description: Date YYYY-MM-DD (UTC day) when lease must be active
        example: "2026-04-18"
        in: query
        name: active_on
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope-leases_LeaseListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: List leases
      tags:
      - leases
    post:
      consumes:
      - application/json
      description: '**Draft** is allowed. **Active + primary** leases cannot overlap
        in time on the same unit (enforced here and by SQL EXCLUDE). Tenant and unit
        must belong to the organization. When `status` is **active**, the unit becomes
        **occupied** if at least one active lease exists. **Audit:** `lease.created`
        is recorded.'
      parameters:
      - description: Organization ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: 'Example: {\'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/leases.CreateLeaseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Envelope-leases_LeaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "409":
          description: lease_overlap
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Create lease
      tags:
      - leases
...
  /api/v1/organizations/{id}/tenants:
    get:
      description: '**Nested route:** `/api/v1/organizations/{id}/tenants`. **Admin:**
        any org `id`. **Landlord, manager, staff:** own organization only. **Pagination:**
        `page`, `page_size` (max 100). **Search (AND):** optional `name`, `email`,
        `phone` query params — each matches case-insensitively (`ILIKE`) against name
        fields or email/phone respectively. Optional `status` filter (`active`, `inactive`,
        `archived`). Set `include_summary=true` to add `active_lease_count` and `current_unit`
        (from the latest active lease) per tenant.'
      parameters:
      - description: Organization ID (Gin param name `id`)
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - default: 1
        description: Page (1-based)
        example: 1
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Page size
        example: 20
        in: query
        maximum: 100
        name: page_size
        type: integer
      - description: Search by first/last/full name (partial)
        example: Jordan
        in: query
        name: name
        type: string
      - description: Search by email (partial)
        example: lee@
        in: query
        name: email
        type: string
      - description: Search by phone (partial)
        example: "555"
        in: query
        name: phone
        type: string
      - description: Filter by tenant status
        example: active
        in: query
        name: status
        type: string
      - description: true/1/yes to include lease summary fields
        example: "true"
        in: query
        name: include_summary
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'Example item: full_name, email_opt_in, active_lease_count
            when include_summary=true'
          schema:
            $ref: '#/definitions/response.Envelope-tenants_TenantListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: List tenants
      tags:
      - tenants
    post:
      consumes:
      - application/json
      description: '**Admin / landlord / manager** only. Validates `email` format
        when provided; `phone` max length 32. If `full_name` is omitted it is derived
        from first + last name. Default `status` is `active`.'
      parameters:
      - description: Organization ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: 'Example: {\'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/tenants.CreateTenantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Envelope-tenants_TenantResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Create tenant
      tags:
      - tenants
...
  /api/v1/organizations/{id}/users:
    get:
      description: '**Admin:** any organization `id`. **Landlord / manager:** only
        users in their own organization (`id` must match the token). **Staff** is
        not allowed. Paginated with `page` and `page_size` (max 100).'
      parameters:
      - description: Organization ID
        example: 550e8400-e29b-41d4-a716-446655440000
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - default: 1
        description: Page number (1-based)
        example: 1
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Page size
        example: 20
        in: query
        maximum: 100
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope-users_UserListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: List organization users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: '**Admin:** may create users in any organization with any role.
        **Landlord:** may create `manager` or `staff` in their org. **Manager:** may
        create `staff` only. Password is required unless `status` is `invited` (server
        generates a temporary password). Email must be unique per organization.'
      parameters:
      - description: Organization ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Create payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Envelope-users_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Create organization user
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"